Remove unused and unnecessary data timings stuff
