+++ src/EnergyPlus/DataTimings.hh
@@ -48,93 +48,14 @@
 #ifndef DataTimings_hh_INCLUDED
 #define DataTimings_hh_INCLUDED
 
-// ObjexxFCL Headers
-#include <ObjexxFCL/Array1D.hh>
-#include <ObjexxFCL/Optional.hh>
-
 // EnergyPlus Headers
 #include <EnergyPlus/Data/BaseData.hh>
 #include <EnergyPlus/EnergyPlus.hh>
 
 namespace EnergyPlus {
 
-// Forward declarations
-struct EnergyPlusData;
-
-#ifdef EP_NO_Timings
-#undef EP_Timings
-#endif
-
-namespace DataTimings {
-
-    // Data
-    // -only module should be available to other modules and routines.
-    // Thus, all variables in this module must be PUBLIC.
-
-    // MODULE PARAMETER DEFINITIONS:
-    int constexpr MaxTimingStringLength()
-    {
-        return 250;
-    } // string length for timing string array
-
-    // Following for calls to routines
-
-    // Types
-
-    struct timings
-    {
-        // Members
-        std::string Element;
-        Real64 rstartTime;
-        Real64 currentTimeSum;
-        int calls;
-
-        // Default Constructor
-        timings() : rstartTime(0.0), currentTimeSum(0.0), calls(0)
-        {
-        }
-    };
-
-    // Functions
-
-    void epStartTime(std::string const &ctimingElementstring);
-
-    void epStopTime(std::string const &ctimingElementstring,
-                    Optional_bool_const printit = _, // true if it should be printed here.
-                    Optional_string_const wprint = _ // only needed (and assumed, if printit is true)
-    );
-
-    void epSummaryTimes(Real64 &TimeUsed_CPUTime);
-
-    Real64 epGetTimeUsed(EnergyPlusData &state, std::string const &ctimingElementstring);
-
-    Real64 epGetTimeUsedperCall(EnergyPlusData &state, std::string const &ctimingElementstring);
-
-    Real64 eptime(EnergyPlusData &state);
-
-    Real64 epElapsedTime();
-
-} // namespace DataTimings
-
 struct DataTimingsData : BaseGlobalStruct
 {
-
-    int NumTimingElements = 0;
-    int MaxTimingElements = 0;
-    Real64 dailyWeatherTime = 0.0;
-    Real64 dailyExteriorEnergyUseTime = 0.0;
-    Real64 dailyHeatBalanceTime = 0.0;
-    Real64 hbdailyInit = 0.0;
-    Real64 hbdailyOutSurf = 0.0;
-    Real64 hbdailyInSurf = 0.0;
-    Real64 hbdailyHVAC = 0.0;
-    Real64 hbdailyRep = 0.0;
-    Real64 clockrate = 0.0;
-    bool lprocessingInputTiming = false;
-    bool lmanageSimulationTiming = false;
-    bool lcloseoutReportingTiming = false;
-    Array1D<DataTimings::timings> Timing;
-
     // Following for calls to routines
 #ifdef EP_Count_Calls
     int NumShadow_Calls = 0;
@@ -157,40 +78,7 @@ struct DataTimingsData : BaseGlobalStruct
 
     void clear_state() override
     {
-        this->NumTimingElements = 0;
-        this->MaxTimingElements = 0;
-        this->dailyWeatherTime = 0.0;
-        this->dailyExteriorEnergyUseTime = 0.0;
-        this->dailyHeatBalanceTime = 0.0;
-        this->hbdailyInit = 0.0;
-        this->hbdailyOutSurf = 0.0;
-        this->hbdailyInSurf = 0.0;
-        this->hbdailyHVAC = 0.0;
-        this->hbdailyRep = 0.0;
-        this->clockrate = 0.0;
-        this->lprocessingInputTiming = false;
-        this->lmanageSimulationTiming = false;
-        this->lcloseoutReportingTiming = false;
-        this->Timing.deallocate();
-
-#ifdef EP_Count_Calls
-        this->NumShadow_Calls = 0;
-        this->NumShadowAtTS_Calls = 0;
-        this->NumClipPoly_Calls = 0;
-        this->NumInitSolar_Calls = 0;
-        this->NumAnisoSky_Calls = 0;
-        this->NumDetPolyOverlap_Calls = 0;
-        this->NumCalcPerSolBeam_Calls = 0;
-        this->NumDetShadowCombs_Calls = 0;
-        this->NumIntSolarDist_Calls = 0;
-        this->NumIntRadExchange_Calls = 0;
-        this->NumIntRadExchangeZ_Calls = 0;
-        this->NumIntRadExchangeMain_Calls = 0;
-        this->NumIntRadExchangeOSurf_Calls = 0;
-        this->NumIntRadExchangeISurf_Calls = 0;
-        this->NumMaxInsideSurfIterations = 0;
-        this->NumCalcScriptF_Calls = 0;
-#endif
+        *this = DataTimingsData();
     }
 };
