+++ src/EnergyPlus/UtilityRoutines.cc
@@ -608,6 +608,39 @@ namespace UtilityRoutines {
         return FuelTypeErrorsFound;
     }
 
+    Real64 epElapsedTime()
+    {
+
+        // FUNCTION INFORMATION:
+        //       AUTHOR         Linda Lawrie
+        //       DATE WRITTEN   February 2012
+        //       MODIFIED       na
+        //       RE-ENGINEERED  na
+
+        // PURPOSE OF THIS FUNCTION:
+        // An alternative method for timing elapsed times is to call the standard
+        // Date_And_Time routine and set the "time".
+
+        // Return value
+        Real64 calctime; // calculated time based on hrs, minutes, seconds, milliseconds
+
+        // FUNCTION LOCAL VARIABLE DECLARATIONS:
+        Array1D<Int32> clockvalues(8);
+        // value(1)   Current year
+        // value(2)   Current month
+        // value(3)   Current day
+        // value(4)   Time difference with respect to UTC in minutes (0-59)
+        // value(5)   Hour of the day (0-23)
+        // value(6)   Minutes (0-59)
+        // value(7)   Seconds (0-59)
+        // value(8)   Milliseconds (0-999)
+
+        date_and_time(_, _, _, clockvalues);
+        calctime = clockvalues(5) * 3600.0 + clockvalues(6) * 60.0 + clockvalues(7) + clockvalues(8) / 1000.0;
+
+        return calctime;
+    }
+
 } // namespace UtilityRoutines
 
 int AbortEnergyPlus(EnergyPlusData &state)
@@ -629,7 +662,6 @@ int AbortEnergyPlus(EnergyPlusData &state)
 
     // Using/Aliasing
     using namespace DataSystemVariables;
-    using namespace DataTimings;
     using namespace DataErrorTracking;
     using BranchInputManager::TestBranchIntegrity;
     using BranchNodeConnections::CheckNodeConnections;
@@ -711,12 +743,9 @@ int AbortEnergyPlus(EnergyPlusData &state)
     NumSevereDuringSizing = fmt::to_string(state.dataErrTracking->TotalSevereErrorsDuringSizing);
 
     // catch up with timings if in middle
-    state.dataSysVars->Time_Finish = epElapsedTime();
+    state.dataSysVars->Time_Finish = UtilityRoutines::epElapsedTime();
     if (state.dataSysVars->Time_Finish < state.dataSysVars->Time_Start) state.dataSysVars->Time_Finish += 24.0 * 3600.0;
     state.dataSysVars->Elapsed_Time = state.dataSysVars->Time_Finish - state.dataSysVars->Time_Start;
-#ifdef EP_Detailed_Timings
-    epStopTime("EntireRun=");
-#endif
     if (state.dataSysVars->Elapsed_Time < 0.0) state.dataSysVars->Elapsed_Time = 0.0;
     Hours = state.dataSysVars->Elapsed_Time / 3600.0;
     state.dataSysVars->Elapsed_Time -= Hours * 3600.0;
@@ -753,9 +782,6 @@ int AbortEnergyPlus(EnergyPlusData &state)
 
     state.dataResultsFramework->resultsFramework->writeOutputs(state);
 
-#ifdef EP_Detailed_Timings
-    epSummaryTimes(state.files.audit, Time_Finish - Time_Start);
-#endif
     std::cerr << "Program terminated: "
               << "EnergyPlus Terminated--Error(s) Detected." << std::endl;
     // Close the socket used by ExternalInterface. This call also sends the flag "-1" to the ExternalInterface,
@@ -817,7 +843,6 @@ int EndEnergyPlus(EnergyPlusData &state)
     // Stops the program.
 
     using namespace DataSystemVariables;
-    using namespace DataTimings;
     using namespace DataErrorTracking;
     using ExternalInterface::CloseSocket;
 
@@ -854,15 +879,12 @@ int EndEnergyPlus(EnergyPlusData &state)
     NumSevereDuringSizing = fmt::to_string(state.dataErrTracking->TotalSevereErrorsDuringSizing);
     strip(NumSevereDuringSizing);
 
-    state.dataSysVars->Time_Finish = epElapsedTime();
+    state.dataSysVars->Time_Finish = UtilityRoutines::epElapsedTime();
     if (state.dataSysVars->Time_Finish < state.dataSysVars->Time_Start) state.dataSysVars->Time_Finish += 24.0 * 3600.0;
     state.dataSysVars->Elapsed_Time = state.dataSysVars->Time_Finish - state.dataSysVars->Time_Start;
     if (state.dataGlobal->createPerfLog) {
         UtilityRoutines::appendPerfLog(state, "Run Time [seconds]", format("{:.2R}", state.dataSysVars->Elapsed_Time));
     }
-#ifdef EP_Detailed_Timings
-    epStopTime("EntireRun=");
-#endif
     Hours = state.dataSysVars->Elapsed_Time / 3600.0;
     state.dataSysVars->Elapsed_Time -= Hours * 3600.0;
     Minutes = state.dataSysVars->Elapsed_Time / 60.0;
@@ -900,9 +922,6 @@ int EndEnergyPlus(EnergyPlusData &state)
 
     state.dataResultsFramework->resultsFramework->writeOutputs(state);
 
-#ifdef EP_Detailed_Timings
-    epSummaryTimes(Time_Finish - Time_Start);
-#endif
     if (state.dataGlobal->printConsoleOutput) std::cerr << "EnergyPlus Completed Successfully." << std::endl;
     // Close the ExternalInterface socket. This call also sends the flag "1" to the ExternalInterface,
     // indicating that E+ finished its simulation
